pyproject.toml
[build-system]
requires = ["scikit-build-core[pyproject]", "numpy"]
build-backend = "scikit_build_core.build"

[project]
name = "pmm_planner"
version = "0.0.1"
description = "PMM Planner"
readme = "README.md"
requires-python = ">=3.10"

dependencies = [
    "numpy >= 2.0.0",
    "matplotlib", 
    "pyyaml", 
    "pybind11-global",
    "toml >= 0.10.2",
    "packaging >= 24.0",
]

[tool.scikit-build]
wheel.packages = ["pmm_planner"]  # <-- THIS IS CRUCIAL
cmake.verbose = true

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".git",
    ".git-rewrite",
    ".ipynb_checkpoints",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "build",
    "dist",
    "site-packages",
    "venv",
    "docs/conf.py",
]

line-length = 100
indent-width = 4
target-version = "py311"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F", "I", "D", "TCH", "ANN"]
ignore = ["ANN401"]
fixable = ["ALL"]
unfixable = []

[tool.ruff.lint.per-file-ignores]
"benchmarks/*" = ["D100", "D103"]
"tests/*" = ["D100", "D103", "D104"]


[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.flake8-annotations]
suppress-dummy-args = true
suppress-none-returning = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = true
line-ending = "auto"
docstring-code-format = true
docstring-code-line-length = "dynamic"
